change train datatype

+++ train.py
@@ -121,7 +121,7 @@ min_lr = learning_rate/10 # minimum learning rate, should be ~= learning_rate/10
 backend = 'nccl' # 'nccl', 'gloo', etc.
 # system
 #device = 'cuda' # examples: 'cpu', 'cuda', 'cuda:0', 'cuda:1' etc., or try 'mps' on macbooks
-dtype = 'bfloat16' # 'float32', 'bfloat16', or 'float16', the latter will auto implement a GradScaler
+dtype = 'float16' # 'float32', 'bfloat16', or 'float16', the latter will auto implement a GradScaler
 compile = True # use PyTorch 2.0 to compile the model to be faster
 
 '''check_type = 'shortest'
